+++ newton/_src/solvers/phoenx/clustering/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 The Newton Developers
-# SPDX-License-Identifier: Apache-2.0
-
-from newton._src.solvers.phoenx.clustering.cluster_builder import ConstraintClusterBuilder
-from newton._src.solvers.phoenx.clustering.clustering_pipeline import ClusteringPipeline
-from newton._src.solvers.phoenx.clustering.supernodal_elements import SupernodalElements
-
-__all__ = ["ClusteringPipeline", "ConstraintClusterBuilder", "SupernodalElements"]